feat: Add bcrypt library for password hashing and reset token functionality
This commit adds the bcrypt library to the project's dependencies in the package.json file. The bcrypt library is used for password hashing and will enhance the security of user passwords. Additionally, the user model in the backend is updated to include a new field, 'resetToken', which will be used for password reset functionality. These changes improve the security and functionality of the application.